uci-defaults: enforce a way to make pbr manual only.
this will remove /etc/rc.d/S94pbr from each upgrade to disable its service automaticly as a work around for the 10+ interfaces which need to reload and causing major issues.

by removing this file, we get the same behaviour as when we disable the service inside luci->system->startup which is not the same result as /etc/init.d/pbr disable.

now the user has to manual press restart in order to do the policy routing/kill switching without causing to much of an issue.

this might be a bug, or a known limitation for pbr this kind of fixes the issue.