[packages]: add ebtables for multicast filtering
Currently when using Avahi we cannot just use ``allow-interfaces`` only ``deny-interfaces`` work,
however the issue is that we now have multicast propogating on networks we absolutely don't want this.

We might aswell create a luci-app-multicast-manager or something because it
is very usefull to drop unsolicitated multicast traffic.

Example:

JetKVM has a feature request, the pc connected to the kvm is on vlan 49 (pcnet) and kvm on vlan 180 (kvmnet),
we do not want JetKVM sending multicast advertisements which make the pc findable... wether due to spotify
or whatever other different traffic.

We want to isolate multicasts per interfaces if needed, and hope Ebtables can do this job.