[profile]: as for now disable ucode-wifi-scripts
as for now temporary disable ucode-wifi-scripts so we use wifi-scripts again.
this is due to a regression in where our exotic wireless configuration breaks.
and the wifi clients cannot connect.

as for now uncheck this option, and only allow it for debugging for when it is fixed.
we need to also migrate as soon as possible to the ucode version, but for now it is incompatible.

The issue has been reported by me to the maintainer on the OpenWrt forums:
https://forum.openwrt.org/t/is-ucode-wifi-scripts-good-to-use-already/222615/9?u=xize