profiles: add privoxy profile
This profile is the same as central with the exception we include privoxy.

Since we do not know if it even compiles, it seems to be not present in OpenWrt snapshot tree except for the luci app, seperate this profile because we think this has a high ability to break.